Adding test for getTodoListTitleDate
